Bundle libappindicator in AppImage & Zip packages
Bundle the system's libappindicator in AppImage & Zip packages.
libappindicator is only required for Gnome desktop. deb and rpm packages
do not use the bundled libappindicator but instead use the 'depends'
mechanism. If libappindicator is not present, the packaging continues
and an info message displayed. If packaging for a target architecture
different than the host architecture, libappindicator is not bundled,
the packaging continues and an info message displayed.