fix(nightly): Don't remove prerelease info
This commit reverts a change that was introduced to remove the
prerelease info before setting the package version to nightly. This is
important because by removing prerelease info, 2.0.0-beta+nightly would
become 2.0.0+nightly, i.e. it would be valued *higher* than the full 2.0
release that will be coming in the future. The build info must always
just be appended to whatever the package version is set to.