feat(file-manager): Adds filter to the file tree
This commit adds the filter functionality from the file list also to the
file tree. This means that from now on you can not only filter the files
in the file list, but additionally the full directory tree, which comes
in especially handy for users who like to use the combined file manager
mode.