feat(FSAL): Better file generation workflow
This commit changes the way Zettlr creates new files. When the setting
newFileDontPrompt is true, it will always create a file on the hard disk
without asking anything -- both when creating a new file from a context
menu as well as when using the shortcut. When that setting is turned
off, however, it will always ask for a filename (rather: a confirmation)
when creating a file using the context menu. When the user uses the
shortcut, however, it will create an in-memory file and only ask for the
filename when the file is being saved.

Fixes #2444, fixes #2310